[SHIVANI] : added Ability to find maximum in given float

+++ src/com/bridgelab/generics/FindMaximum.java
@@ -3,12 +3,11 @@ package com.bridgelab.generics;
 public class FindMaximum {
 
 	/*
-	 * UC1 - Find maximum in given three integer
+	 * UC2 - Find maximum in given three float
 	 */
-	
-	public Integer checkMaximum(Integer X, Integer Y, Integer Z) {
+	public Float checkMaximum(Float X, Float Y, Float Z) {
 		
-		Integer max = X;
+		Float max = X;
 		if(Y.compareTo(max) > 0) {
 			max = Y;
 		}
